feat: add product entity, validation, and soft delete functionality
- Introduced `Product` entity with validation and serialization groups
- Added soft delete functionality to `StatisticsPropertiesTraits`
- Updated `Farm` entity to include product relationships and validation
- Enhanced `FarmController` to handle product associations and validation
- Added `ExceptionSubscriber` for handling HTTP exceptions
- Updated dependencies to include `symfony/validator` and `doctrine/annotations`

+++ src/Controller/FarmController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Farm;
 use App\Repository\FarmRepository;
+use App\Repository\ProductRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -13,6 +14,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 final class FarmController extends AbstractController
 {
@@ -20,21 +22,33 @@ final class FarmController extends AbstractController
     public function getAll(FarmRepository $farmFarmRepository, SerializerInterface $serializer): JsonResponse
     {
         $farms = $farmFarmRepository->findAll();
-        $jsonData = $serializer->serialize($farms, 'json');
+        $jsonData = $serializer->serialize($farms, 'json', ['groups' => ['farm', 'stats']]);
         return new JsonResponse($jsonData, Response::HTTP_OK, [], true);
     }
 
     #[Route('api/v1/farm/{farm}', name: 'get_farm', methods: ['GET'])]
     public function get(Farm $farm, SerializerInterface $serializer): JsonResponse
     {
-        $jsonData = $serializer->serialize($farm, 'json');
+        $jsonData = $serializer->serialize($farm, 'json', ['groups' => ['farm','stats']]);
         return new JsonResponse($jsonData, Response::HTTP_OK, [], true);
     }
 
     #[Route('api/v1/farm', name: 'create_farm', methods: ['POST'])]
-    public function create(Request $request, UrlGeneratorInterface $urlGenerator, SerializerInterface $serializer, EntityManagerInterface $entityManager): JsonResponse
+    public function create(Request $request, UrlGeneratorInterface $urlGenerator, ProductRepository $productRepository, SerializerInterface $serializer, EntityManagerInterface $entityManager, ValidatorInterface $validator): JsonResponse
     {
         $farm = $serializer->deserialize($request->getContent(), Farm::class, 'json');
+        $productsData = $request->toArray()['products'] ?? [];
+        foreach ($productsData as $productId) {
+            $product = $productRepository->find($productId);
+            if ($product) {
+                $farm->addProduct($product);
+            }
+        }
+        $farm->setStatus('on');
+        $errors = $validator->validate($farm);
+        if (count($errors) > 0) {
+            return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+        }
         $entityManager->persist($farm);
         $entityManager->flush();
         $jsonData = $serializer->serialize($farm, 'json');
@@ -43,11 +57,24 @@ final class FarmController extends AbstractController
     }
 
     #[Route("api/v1/farm/{farm}", name:"update_farm", methods: ["PATCH"])]
-    public function update(Farm $farm, Request $request, SerializerInterface $serializer, EntityManagerInterface $entityManager): JsonResponse 
-    { 
+    public function update(Farm $farm, Request $request, ProductRepository $productRepository, SerializerInterface $serializer, EntityManagerInterface $entityManager, ValidatorInterface $validator): JsonResponse 
+    {
         $updatedFarm = $serializer->deserialize(
             $request->getContent(), Farm::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $farm]
         );
+        
+        $productsData = $request->toArray()['products'] ?? [];
+        $farm->getProducts()->clear();
+        foreach ($productsData as $productId) {
+            $product = $productRepository->find($productId);
+            if ($product) {
+                $farm->addProduct($product);
+            }
+        }
+        $errors = $validator->validate($updatedFarm);
+        if (count($errors) > 0) {
+            return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+        }
         
         $entityManager->persist($updatedFarm);
         $entityManager->flush();
@@ -55,9 +82,16 @@ final class FarmController extends AbstractController
     }
 
     #[Route("api/v1/farm/{farm}", name:"delete_farm", methods: ["DELETE"])]
-    public function delete(Farm $farm, EntityManagerInterface $entityManager): JsonResponse 
-    { 
-        $entityManager->remove($farm);
+    public function delete(Farm $farm, Request $request, EntityManagerInterface $entityManager): JsonResponse 
+    {
+        $requestData = json_decode($request->getContent(), true);
+        
+        if(isset($requestData['hard']) && $requestData['hard'] === true) {
+            $entityManager->remove($farm);
+        } else {
+            $farm->softDelete();
+        }
+        
         $entityManager->flush();
         return new JsonResponse(null, Response::HTTP_NO_CONTENT, []); 
     }
